fix(ui) : implement installSplashScreen
- Add Splash API library to implement android splash
- Delete SplashFragment.kt+xml as the file/fragment is currently not used anymore
