+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,8 @@
             android:exported="false" />
         <activity
             android:name=".presentation.MainActivity"
-            android:exported="true">
+            android:exported="true"
+            android:theme="@style/Theme.Satakomobile.Splash">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
