+++ app/src/main/java/dev/capstone/satako_mobile/presentation/MainActivity.kt
@@ -5,10 +5,13 @@ import android.util.Log
 import android.view.View
 import android.window.OnBackInvokedDispatcher
 import androidx.activity.enableEdgeToEdge
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
 import androidx.navigation.NavController
 import androidx.navigation.findNavController
 import androidx.navigation.fragment.NavHostFragment
@@ -18,17 +21,32 @@ import androidx.navigation.ui.setupWithNavController
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import dev.capstone.satako_mobile.R
 import dev.capstone.satako_mobile.databinding.ActivityMainBinding
+import dev.capstone.satako_mobile.presentation.splash.SplashViewModel
 import dev.capstone.satako_mobile.utils.gone
 import dev.capstone.satako_mobile.utils.show
 
 class MainActivity : AppCompatActivity() {
 
+    private val splashViewModel: SplashViewModel by viewModels {
+        ViewModelFactory(this)
+    }
     private lateinit var binding: ActivityMainBinding
     private lateinit var navController: NavController
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        installSplashScreen()
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
+        val navHostFragment =
+            supportFragmentManager.findFragmentById(R.id.nav_host_fragment_activity_main) as NavHostFragment
+        val navController = navHostFragment.navController
+        val navGraph = navController.navInflater.inflate(R.navigation.main_navigation)
+        splashViewModel.getSession().observe(this) { session ->
+            navGraph.setStartDestination(
+                if (session.isNotEmpty()) R.id.home_fragment else R.id.onboardingFragment
+            )
+            navController.graph = navGraph
+        }
         navigationBar()
     }
