+++ app/src/main/java/dev/capstone/satako_mobile/presentation/splash/PlaceholderFragment.kt
@@ -1,0 +1,23 @@
+package dev.capstone.satako_mobile.presentation.splash
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import dev.capstone.satako_mobile.R
+class PlaceholderFragment : Fragment() {
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_placeholder, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+    }
+}
