+++ app/src/main/res/drawable/icon_splash.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="74dp"
+    android:height="74dp"
+    android:viewportWidth="74"
+    android:viewportHeight="74">
+  <path
+      android:pathData="M37,0C57.435,0 74,16.566 74,37C74,57.435 57.435,74 37,74C27.521,74 18.875,70.436 12.328,64.574C12.328,64.574 12.329,64.574 12.329,64.574C13.532,63.95 14.411,63.298 15.223,62.695C16.949,61.415 18.371,60.361 21.948,60.278C23.435,60.243 25.81,60.442 28.489,60.667C31.786,60.943 35.544,61.258 38.676,61.224C43.495,61.171 49.181,55.637 53.112,51.81C53.654,51.282 54.162,50.787 54.631,50.343L63.199,41.303C65.613,38.414 62.1,33.505 58.313,37.306L41.229,51.208C39.006,52.898 38.299,52.882 35.816,52.825C35.296,52.813 34.697,52.8 33.99,52.8C33.521,52.8 33.051,52.799 32.582,52.799C31.564,52.797 30.547,52.796 29.53,52.801C29.27,52.802 29.016,52.815 28.774,52.827C27.605,52.884 26.716,52.928 26.75,51.629C27.653,50.829 29.902,50.984 32.33,51.151C35.907,51.397 39.874,51.67 40.507,48.96C41.414,45.071 36.992,45.187 32.188,45.313C30.77,45.35 29.319,45.388 27.962,45.325C26.934,45.276 26.013,45.121 25.104,44.969C23.642,44.723 22.213,44.483 20.425,44.699C16.295,45.199 14.083,47.128 12.246,48.73C11.406,49.463 10.644,50.128 9.812,50.556C8.39,51.288 5.037,52.883 3.755,53.26C3.755,53.26 3.755,53.26 3.755,53.261C1.35,48.353 0,42.834 0,37C0,16.566 16.566,0 37,0ZM6.873,58.484C7.428,59.261 8.013,60.017 8.626,60.748C8.034,60.028 7.439,59.264 6.887,58.504C6.882,58.497 6.877,58.491 6.873,58.484ZM42.306,48.361C41.838,46.452 41.788,46.368 40.989,45.089L47.823,39.974C47.969,39.85 48.098,39.742 48.218,39.642C48.638,39.29 48.95,39.029 49.485,38.506C51.317,36.717 52.87,35.84 55.333,37.508C55.167,38.108 53.84,39.161 52.933,39.881C52.665,40.094 52.434,40.278 52.28,40.412C50.319,42.122 48.303,43.677 46.3,45.221C44.955,46.258 43.616,47.29 42.306,48.361ZM24.97,38.132C23.177,39.427 15.08,37.666 12.358,36.363C12.199,35.717 12.11,35.34 12.271,34.537C16.009,36.425 18.424,36.839 20.9,37.263C22.178,37.482 23.471,37.703 24.97,38.132ZM41.849,37.689C41.996,37.802 42.108,37.887 42.174,37.932C42.117,37.963 42.082,37.983 42.056,37.999C41.997,38.033 41.984,38.041 41.838,38.096C40.938,38.435 35.723,39.475 35.947,36.957C36.261,33.43 40.52,36.676 41.849,37.689ZM31.358,37.918C32.635,36.259 34.848,38.834 35.722,39.892L36.691,41.384C35.111,41.619 29.302,40.588 31.358,37.918ZM38.659,25.785C38.528,25.363 38.427,25.039 38.435,24.483C39.651,25.028 41.31,25.832 42.199,26.702C44.679,29.127 40.814,31.739 39.407,27.856C39.314,27.6 39.231,27.407 39.152,27.224C39.028,26.935 38.913,26.671 38.789,26.221C38.744,26.059 38.7,25.917 38.659,25.785ZM26.645,32.548C26.987,32.188 27.324,31.833 27.619,31.422C28.238,32.929 28.465,38.297 26.028,37.652C24.379,37.215 24.478,35.774 25.185,34.434C25.598,33.651 26.127,33.094 26.645,32.548ZM31.404,30.689C31.494,30.557 31.574,30.44 31.639,30.341C31.778,30.527 31.798,30.538 31.798,30.538C31.797,30.537 31.773,30.522 31.858,30.71L32.166,32.03C32.291,32.65 32.364,33.047 32.423,33.691C32.526,34.823 32.396,36.768 30.987,36.737C27.349,36.655 30.244,32.396 31.404,30.689ZM43.02,19.636C43.023,18.339 43.535,16.818 44.065,15.496C44.903,16.359 47.735,20.818 45.405,21.6C43.822,22.132 43.018,20.939 43.02,19.636ZM50.464,20.691C49.446,20.737 44.183,22.325 45.423,24.542C47.299,27.898 49.629,22.713 50.464,20.691ZM46.22,32.735C46.33,32.77 46.44,32.805 46.549,32.84C46.061,33.435 44.251,34.445 43.631,34.762C39.863,36.683 38.897,31.926 42.458,31.915C43.664,31.911 44.971,32.332 46.22,32.735ZM49.443,26.81C47.721,26.669 42.698,26.999 43.877,29.671C45.179,32.624 48.425,28.751 49.443,26.81ZM36.093,35.337C34.792,36.433 32.054,35.05 34.242,30.336L34.77,29.198C35.252,29.639 37.662,34.015 36.093,35.337ZM36.674,30.991C36.517,29.745 36.685,28.236 37.018,26.913C38.384,28.131 41.729,31.957 39.052,33.093C37.439,33.777 36.808,32.05 36.674,30.991ZM39.18,21.021C39.409,21.9 42.074,26.919 44.024,25.132C46.92,22.479 40.21,20.763 39.18,21.021ZM30.661,42.045C30.794,42.356 30.93,42.674 31.101,43.022C28.955,42.663 23.341,40.864 26.377,38.453C27.515,37.55 28.937,39.045 29.559,39.955C30.117,40.771 30.383,41.394 30.661,42.045Z"
+      android:fillColor="@color/primary"
+      android:fillType="evenOdd"/>
+</vector>
