+++ app/src/main/res/drawable/splash_screen_dark.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@color/background_dark" />
+    <item
+        android:width="150dp"
+        android:height="150dp"
+        android:drawable="@drawable/icon_splash_dark"
+        android:gravity="center" />
+</layer-list>
