+++ app/src/main/res/layout/fragment_login.xml
@@ -156,7 +156,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/or_tv" />
-    
+
 
     <ProgressBar
         android:id="@+id/pb_login"
