+++ app/src/main/res/navigation/main_navigation.xml
@@ -3,31 +3,8 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main_navigation"
-    app:startDestination="@id/splashFragment">
+    app:startDestination="@id/placeholderFragment2">
 
-    <fragment
-        android:id="@+id/splashFragment"
-        android:name="dev.capstone.satako_mobile.presentation.splash.SplashFragment"
-        android:label="fragment_splash"
-        tools:layout="@layout/fragment_splash">
-        <action
-            android:id="@+id/action_splashFragment_to_onboardingFragment"
-            app:destination="@id/onboardingFragment"
-            app:enterAnim="@anim/nav_default_enter_anim"
-            app:exitAnim="@anim/nav_default_exit_anim"
-            app:launchSingleTop="true"
-            app:popUpTo="@id/splashFragment"
-            app:popUpToInclusive="true" />
-        <action
-            android:id="@+id/action_splashFragment_to_homeFragment"
-            app:destination="@id/home_fragment"
-            app:enterAnim="@anim/from_right"
-            app:exitAnim="@anim/to_left"
-            app:launchSingleTop="true"
-            app:popUpTo="@id/splashFragment"
-            app:popUpToInclusive="true" />
-
-    </fragment>
     <fragment
         android:id="@+id/onboardingFragment"
         android:name="dev.capstone.satako_mobile.presentation.onboarding.OnboardingFragment"
@@ -182,4 +159,9 @@
             android:name="article_item"
             app:argType="dev.capstone.satako_mobile.data.model.dummy.Article" />
     </fragment>
+    <fragment
+        android:id="@+id/placeholderFragment2"
+        android:name="dev.capstone.satako_mobile.presentation.splash.PlaceholderFragment"
+        android:label="fragment_placeholder"
+        tools:layout="@layout/fragment_placeholder" />
 </navigation>
