+++ gradle/libs.versions.toml
@@ -1,6 +1,7 @@
 [versions]
 agp = "8.3.2"
 cameraxVersion = "1.4.0-beta01"
+coreSplashscreen = "1.0.1"
 credentials = "1.5.0-alpha01"
 credentialsPlayServicesAuth = "1.2.2"
 credentialsVersion = "1.2.2"
@@ -36,6 +37,7 @@ androidx-camera-lifecycle = { module = "androidx.camera:camera-lifecycle", versi
 androidx-camera-video = { module = "androidx.camera:camera-video", version.ref = "cameraxVersion" }
 androidx-camera-view = { module = "androidx.camera:camera-view", version.ref = "cameraxVersion" }
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
+androidx-core-splashscreen = { module = "androidx.core:core-splashscreen", version.ref = "coreSplashscreen" }
 androidx-credentials-play-services-auth-v122 = { module = "androidx.credentials:credentials-play-services-auth", version.ref = "credentialsPlayServicesAuth" }
 androidx-credentials-v122 = { module = "androidx.credentials:credentials", version.ref = "credentialsVersion" }
 androidx-datastore-preferences = { module = "androidx.datastore:datastore-preferences", version.ref = "datastoreCoreAndroid" }
